+++ R/z_utils.R
@@ -79,7 +79,8 @@ o2s = function(obj) {
              "surveytable.survey_label", "surveytable.svychisq_statistic",
              "surveytable.svyciprop_adj", "surveytable.tx_count", "surveytable.tx_df",
              "surveytable.tx_numeric", "surveytable.tx_prct", "surveytable.tx_pval",
-             "surveytable.tx_rate", "surveytable.tx_test_stat")
+             "surveytable.tx_rate", "surveytable.tx_test_stat",
+             "surveytable.restructure_attr", "surveytable.restructure_flags")
   c_have = show_options() %>% names()
 
   d_nh = setdiff(c_need, c_have)